Use breakword package
