+++ .isort.cfg
@@ -1,6 +1,6 @@
 [settings]
 known_first_party=myia,debug
-known_third_party=torch,tvm,nnvm,hrepr,buche
+known_third_party=torch,tvm,nnvm,hrepr,buche,breakword
 multi_line_output=3
 include_trailing_comma=True
 combine_as_imports=True
